backend/requirements.txt
firebase-admin>=6.5.0
openai>=1.30.0
python-dotenv>=1.0.1
